requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
gspread
gspread-dataframe
oauth2client
